backend/requirements.txt
fastapi
uvicorn
python-multipart
rembg
pillow
diffusers
transformers
torch
